Rename now that we use a different import style
