+++ dilib/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 from dilib.container import ConfigProxy, Container, get_container
-from dilib.di_config import Config, get_config
+from dilib.config import Config, get_config
 from dilib.errors import (
     ConfigError,
     FrozenConfigError,
